Map whitebox adjustments + audio fixes + lights
Fixed certain parts of the whitebox to better fit the level and the audiosources therein.
Adjusted volumes and falloff variables for various sources. (All apart from the running monster and crying vent monster and for any sources close to the player, ie. footsteps)
Fixed path for wind sound to be linear instead of a curve, allowing it to follow a straight path through the corridors.
Added light sources for playtest debugging so we are aware where the playtester is at any moment.
Added Project Acoustics, baked the map and added various _ATT sources to sound sources to use the plugin.
Still needs debugging and fixing to sort out occlusion to make sounds sound like they're around a corner.